+++ packages/next-common/components/overview/accountInfo/accountInfoPanel.js
@@ -26,8 +26,6 @@ import { isNil } from "lodash-es";
 import Link from "next/link";
 import Button from "next-common/lib/button";
 import AccountPanelQuickAccess from "./components/accountPanelQuickAccess";
-import AccountUnlockBalancePrompt from "./components/accountUnlockBalancePrompt";
-import WithPallet from "next-common/components/common/withPallet";
 import useRealAddress from "next-common/utils/hooks/useRealAddress";
 import Avatar from "next-common/components/avatar";
 import getIpfsLink from "next-common/utils/env/ipfsEndpoint";
@@ -295,7 +293,7 @@ export function AccountHead({ width }) {
   );
 }
 
-export function CommonAccountInfoPanel() {
+export default function AccountInfoPanel() {
   const { width } = useWindowSize();
 
   if (isNil(width)) {
@@ -309,14 +307,7 @@ export function CommonAccountInfoPanel() {
       <Divider />
       <AccountBalances />
       <ExtensionUpdatePrompt />
-      <WithPallet pallet="referenda">
-        <AccountUnlockBalancePrompt />
-      </WithPallet>
       <AccountPanelScrollPrompt />
     </NeutralPanel>
   );
-}
-
-export default function AccountInfoPanel() {
-  return <CommonAccountInfoPanel />;
 }