+++ packages/next-common/components/overview/accountInfo/components/useDelegationPrompt.js
@@ -1,4 +1,8 @@
-import { PromptTypes } from "next-common/components/scrollPrompt";
+import { isEmpty } from "lodash-es";
+import {
+  PromptTypes,
+  ScrollPromptItemWrapper,
+} from "next-common/components/scrollPrompt";
 import { useChain, useChainSettings } from "next-common/context/chain";
 import { isKintsugiChain } from "next-common/utils/chain";
 import { CACHE_KEY } from "next-common/utils/constants";
@@ -39,4 +43,23 @@ export default function useDelegationPrompt() {
       close: () => setVisible(false, { expires: 15 }),
     };
   }, [setVisible, hasDelegation, visible]);
+}
+
+export function DelegationPrompt({ onClose }) {
+  const prompt = useDelegationPrompt();
+  if (isEmpty(prompt)) {
+    return null;
+  }
+
+  return (
+    <ScrollPromptItemWrapper
+      prompt={{
+        ...prompt,
+        close: () => {
+          onClose?.();
+          prompt?.close();
+        },
+      }}
+    />
+  );
 }