+++ packages/next-common/components/preImages/createPreimagePopup/templates/killReferendumPopup.js
@@ -8,12 +8,13 @@ import useReferendumIndexField from "../fields/useReferendumIndexField";
 import { isNil } from "lodash-es";
 import { usePopupParams } from "next-common/components/popupWithSigner/context";
 import InsufficientBalanceTips from "next-common/components/summary/newProposalQuickStart/common/insufficientBalanceTips";
+import ExtrinsicInfo from "../../newPreimagePopup/info";
 
 export function useKillReferendumNotePreimageTx(referendumIndex) {
   const api = useContextApi();
 
   return useMemo(() => {
-    if (!api || isNil(referendumIndex)) {
+    if (!api || isNil(referendumIndex) || referendumIndex === "") {
       return {};
     }
 
@@ -31,13 +32,20 @@ export default function KillReferendumPopup() {
   const { value: referendumIndex, component: referendumIndexField } =
     useReferendumIndexField();
 
-  const { notePreimageTx, encodedLength } =
+  const { notePreimageTx, encodedLength, encodedProposal, encodedHash } =
     useKillReferendumNotePreimageTx(referendumIndex);
 
   return (
     <Popup title="Kill a referendum" onClose={onClose}>
       <SignerWithBalance />
       {referendumIndexField}
+      {encodedProposal && (
+        <ExtrinsicInfo
+          preimageHash={encodedHash}
+          callData={encodedProposal}
+          preimageLength={encodedLength || 0}
+        />
+      )}
       <InsufficientBalanceTips byteLength={encodedLength} onlyPreimage />
       <div className="flex justify-end">
         <NotePreimageButton notePreimageTx={notePreimageTx} />