+++ packages/next-common/components/profile/relatives.jsx
@@ -1,10 +1,9 @@
 import Button from "next-common/lib/button";
 import dynamic from "next/dynamic";
 import { useState } from "react";
-import { isPolkadotChain, isKusamaChain } from "next-common/utils/chain";
-import { useChain } from "next-common/context/chain";
 import dynamicPopup from "next-common/lib/dynamic/popup";
 import useProfileAddress from "./useProfileAddress";
+import RelativesWithNullGuard from "next-common/components/relationshipPopup/relativesWithNullGuard";
 
 const ArrowRight = dynamic(
   import("@osn/icons/subsquare").then((mod) => mod.ArrowRight),
@@ -13,16 +12,6 @@ const SystemRelatives = dynamic(
   import("@osn/icons/subsquare").then((mod) => mod.SystemRelatives),
 );
 const RelativesPopup = dynamicPopup(() => import("../relationshipPopup"));
-
-function RelativesWithNullGuard({ children }) {
-  const chain = useChain();
-
-  if (!isPolkadotChain(chain) && !isKusamaChain(chain)) {
-    return null;
-  }
-
-  return children;
-}
 
 export default function Relatives() {
   const address = useProfileAddress();