+++ packages/next-common/components/relationshipPopup/statusEdge.jsx
@@ -4,11 +4,16 @@ import {
   getSmoothStepPath,
   useNodesData,
 } from "@xyflow/react";
+import { isNil } from "lodash-es";
 import { DisplayUser } from "next-common/components/profile/bio";
-import { indications } from "next-common/components/relationshipPopup/indications";
+import { allIndications } from "next-common/components/relationshipPopup/indications";
 import Tooltip from "next-common/components/tooltip";
-import { rootNodeId } from "next-common/hooks/useConversionRelationshipNode";
+import { rootNodeId } from "next-common/hooks/useRelationshipNode";
 import styled from "styled-components";
+import { useTrackContent } from "../referenda/track/trackTag";
+import ValueDisplay from "../valueDisplay";
+import { toPrecision } from "next-common/utils";
+import { useChainSettings } from "next-common/context/chain";
 
 const EdgeLabel = styled.div`
   position: absolute;
@@ -43,7 +48,7 @@ export default function StatusEdge({
     centerX: edgePathCenterX,
   });
 
-  const edgeTheme = indications.find((item) => item.name === data?.type);
+  const edgeTheme = allIndications.find((item) => item.name === data?.type);
   const sourceNode = useNodesData(source);
   const targetNode = useNodesData(target);
 
@@ -86,6 +91,7 @@ export default function StatusEdge({
                 source={sourceNode?.data?.address}
                 target={targetNode?.data?.address}
                 value={data.value}
+                rawData={data}
               />
             }
           >
@@ -123,6 +129,10 @@ function TooltipsContent({ type, ...rest }) {
 
   if (type === "Identity") {
     return <IdentityTipContent {...rest} />;
+  }
+
+  if (type === "Delegation") {
+    return <DelegationTipContent {...rest} />;
   }
 
   return null;
@@ -167,4 +177,33 @@ function IdentityTipContent({ source, target, value }) {
       <DisplayUser id={target} className="flex text12Medium text-white" />
     </div>
   );
+}
+
+function DelegationTipContent({ rawData }) {
+  const { decimals, symbol } = useChainSettings();
+  if (isNil(rawData) || isNil(rawData.tracks)) {
+    return null;
+  }
+
+  const items = Array.from(rawData.tracks.values());
+
+  return (
+    <ul className="text12Medium">
+      {items.map((track) => (
+        <li key={track.trackId} className="flex items-center">
+          <TrackItem id={track.trackId} />
+          <span className="mr-1">:</span>
+          <ValueDisplay
+            value={toPrecision(track.balance, decimals)}
+            symbol={symbol}
+          />
+        </li>
+      ))}
+    </ul>
+  );
+}
+
+function TrackItem({ id }) {
+  const trackInfo = useTrackContent(id);
+  return <span>{trackInfo}</span>;
 }