+++ packages/next-common/hooks/useDelegatorsRelationshipNode.js
@@ -1,0 +1,100 @@
+import { RELATIONSHIP_NODE_TYPE } from "next-common/utils/constants";
+import pluralize from "pluralize";
+import useMaybeContextMultisigAddress from "./useMaybeContextMultisigAddress";
+import useDelegated, { useDelegators } from "./useRelativesDelegators";
+import {
+  BadgeInfo,
+  DynamicPureProxy,
+  createRootNode,
+  createRelationship,
+  EMPTY_RESULT,
+  rootNodeId,
+} from "./useRelationshipNode";
+
+function createDelegatorsRelationship(rootNode, delegators = []) {
+  return createRelationship({
+    rootNode,
+    items: delegators,
+    nodeIdPrefix: "delegators",
+    edgeIdPrefix: "root-delegators",
+    nodeDataMapper: (item) => ({
+      tracks: item.tracks,
+      address: item.account,
+      badge: <BadgeInfo address={item.account} />,
+      pure: <DynamicPureProxy address={item.account} />,
+    }),
+    edgeDataMapper: (item) => ({
+      type: RELATIONSHIP_NODE_TYPE.Delegation,
+      value: `${item.tracks.size} ${pluralize("track", item.tracks.size)}`,
+      name: "Delegation",
+      tracks: item.tracks,
+    }),
+    sourceKey: "node",
+    targetKey: rootNodeId,
+    sourceHandle: "sourceSub",
+    targetHandle: "targetParent",
+  });
+}
+
+function createDelegatedRelationship(rootNode, delegated = []) {
+  return createRelationship({
+    rootNode,
+    items: delegated,
+    nodeIdPrefix: "delegated",
+    edgeIdPrefix: "root-delegated",
+    nodeDataMapper: (item) => ({
+      tracks: item.tracks,
+      address: item.account,
+      badge: <BadgeInfo address={item.account} />,
+      pure: <DynamicPureProxy address={item.account} />,
+    }),
+    edgeDataMapper: (item) => ({
+      type: RELATIONSHIP_NODE_TYPE.Delegation,
+      value: `${item.tracks.size} ${pluralize("track", item.tracks.size)}`,
+      name: "Delegation",
+      tracks: item.tracks,
+    }),
+    sourceKey: rootNodeId,
+    targetKey: "node",
+    sourceHandle: "sourceSub",
+    targetHandle: "targetParent",
+  });
+}
+
+export default function useDelegatorsRelationshipNode(sourceAddress = "") {
+  const { result: delegators, loading: delegatorsLoading } =
+    useDelegators(sourceAddress);
+
+  const { result: delegated, loading: delegatedLoading } =
+    useDelegated(sourceAddress);
+
+  const isLoading = delegatorsLoading || delegatedLoading;
+
+  const multisigAddress = useMaybeContextMultisigAddress(sourceAddress);
+  const rootNode = createRootNode(sourceAddress, multisigAddress);
+
+  if (rootNode?.data) {
+    rootNode.data.pure = (
+      <DynamicPureProxy
+        address={sourceAddress}
+        className="inline-flex absolute h-5 right-2 top-2"
+      />
+    );
+  }
+
+  if (!sourceAddress) {
+    return EMPTY_RESULT;
+  }
+
+  const { nodes: delegatorsNodes, edges: delegatorsEdges } =
+    createDelegatorsRelationship(rootNode, delegators);
+
+  const { nodes: delegatedNodes, edges: delegatedEdges } =
+    createDelegatedRelationship(rootNode, delegated);
+
+  return {
+    isLoading,
+    nodes: [rootNode, ...delegatorsNodes, ...delegatedNodes],
+    edges: [...delegatorsEdges, ...delegatedEdges],
+  };
+}