+++ packages/next-common/utils/dv/delegates/kusama.js
@@ -2,6 +2,7 @@ import {
   cohort2TrackIds,
   cohort3TrackIds,
   cohort4TrackIds,
+  cohort5TrackIds,
   treasuryTrackIds,
 } from "./common";
 
@@ -88,14 +89,14 @@ export const kusamaDvDelegates = [
     address: "JHTfbt39EL1CcbKteN6hG5L5pWo9XWi9XFiyuS9q24cAc8u", // KusDAO
     slots: [
       { start: 25732465, end: 27921177, trackIds: cohort3TrackIds }, // cohort 3
-      { start: 27921178, end: null, trackIds: cohort4TrackIds }, // cohort 4
+      { start: 27921178, end: 29911969, trackIds: cohort4TrackIds }, // cohort 4
     ],
   },
   {
     address: "Hgm7ELPfRmPKbHgGZCYEZGTjJX8VicXEnFKec7YAeFgAd4d", // Polkadot Hungary
     slots: [
       { start: 25732465, end: 27921177, trackIds: cohort3TrackIds }, // cohort 3
-      { start: 27921178, end: null, trackIds: cohort4TrackIds }, // cohort 4
+      { start: 27921178, end: 29911969, trackIds: cohort4TrackIds }, // cohort 4
     ],
   },
   {
@@ -114,36 +115,113 @@ export const kusamaDvDelegates = [
     address: "HcEbeTviCK33EddVN3mfJ6WymWLyKfFuekjhjn5PFirjJ5F", // SAXEMBERG Governance
     slots: [
       { start: 25732529, end: 27921225, trackIds: cohort3TrackIds }, // cohort 3
+      { start: 29912325, end: null, trackIds: cohort5TrackIds }, // cohort 5
     ],
   },
   {
     address: "HYmYudY1cxN6XyY98dd82TckYF2YiPFc6sXmHqMoKifGAje", // Le Nexus
     slots: [
       { start: 25732529, end: 27921225, trackIds: cohort3TrackIds }, // cohort 3
+      { start: 29912325, end: null, trackIds: cohort5TrackIds }, // cohort 5
     ],
   },
   {
     address: "ELCdsyWFNC7twEeBcQvdpCmpJhGBgiVeWtaKqRqXGn5ATiA", // PERMANENCE DAO/DV
     slots: [
-      { start: 27921280, end: null, trackIds: cohort4TrackIds }, // cohort 4
+      { start: 27921280, end: 29911925, trackIds: cohort4TrackIds }, // cohort 4
+      { start: 29912282, end: null, trackIds: cohort5TrackIds }, // cohort 5
     ],
   },
   {
     address: "EwWrc8UZxaLE8WqCHkygUWAz1PxLc1Jdgzq1kMd8Ac7hKqF", // JAM Implementers DAO
     slots: [
-      { start: 27921529, end: null, trackIds: cohort4TrackIds }, // cohort 4
+      { start: 27921529, end: 29912023, trackIds: cohort4TrackIds }, // cohort 4
     ],
   },
   {
     address: "E3Ra4aGnmZGGtGaLsoCqjtJowT1qDvuLNEwB8t74M1UQrWM", // Trustless Core
     slots: [
-      { start: 27921516, end: null, trackIds: cohort4TrackIds }, // cohort 4
+      { start: 27921516, end: 29912023, trackIds: cohort4TrackIds }, // cohort 4
+      { start: 29912340, end: null, trackIds: cohort5TrackIds }, // cohort 5
     ],
   },
   {
     address: "DaCSCEQBRmMaBLRQQ5y7swdtfRzjcsewVgCCmngeigwLiax", // Polkaworld
     slots: [
-      { start: 27921459, end: null, trackIds: cohort4TrackIds }, // cohort 4
+      { start: 27921459, end: 29911925, trackIds: cohort4TrackIds }, // cohort 4
+    ],
+  },
+  {
+    address: "GykHmXkXiMHV2hnsMdZ7xE7zgd9tiwT8k787MovVavAVmTH", // REEEEEEEEEE DAO
+    slots: [
+      { start: 29912282, end: null, trackIds: cohort5TrackIds }, // cohort 5
+    ],
+  },
+  {
+    address: "DvJWp99ooffSqbTaM3sCYCxLUWbz2eEVqra9oeUZZYbMY14", // PBA Alumni Voting DAO
+    slots: [
+      { start: 29912340, end: null, trackIds: cohort5TrackIds }, // cohort 5
+    ],
+  },
+  {
+    address: "EYSyMJjPk5HJb2ZDAYmEpBu7ZgWm7hZ5b1BE88uCifynRgt", // Polkadot Poland DAO
+    slots: [
+      { start: 29912374, end: null, trackIds: cohort5TrackIds }, // cohort 5
+    ],
+  },
+  {
+    address: "FPznjjQJpHieoy3TUruw9YT6DDRETkBxWv3yFEVUMCgn8q8", // The White Rabbit
+    slots: [
+      {
+        start: 29912374,
+        end: null,
+        trackIds: cohort5TrackIds,
+        role: "guardian",
+      }, // cohort 5
+    ],
+  },
+  {
+    address: "EvoLanodoqDsgHb98Ymbu41uXXKfCPDKxeM6dXHyJ2JoVus", // Daniel Olano
+    slots: [
+      {
+        start: 29912400,
+        end: null,
+        trackIds: cohort5TrackIds,
+        role: "guardian",
+      }, // cohort 5
+    ],
+  },
+  {
+    address: "DuCg7rhST4TX6DWsyePUjntsmJd6UNyQVTHWD5BFjcgmgWp", // Flez
+    slots: [
+      {
+        start: 29912400,
+        end: null,
+        trackIds: cohort5TrackIds,
+        role: "guardian",
+      }, // cohort 5
+    ],
+  },
+  {
+    address: "EyPcJsHXv86Snch8GokZLZyrucug3gK1RAghBD2HxvL1YRZ", // Cybergov — AI Agents(Karim)
+    slots: [
+      {
+        start: 29912423,
+        end: null,
+        trackIds: cohort5TrackIds,
+        role: "guardian",
+      }, // cohort 5
+    ],
+  },
+  {
+    address: "GNdJk9L6P84JXu6wibTzwPiB3vt2rwMjzGEETchf87uNuyW", // GoverNoun AI (Governance Agent) — AI Agent
+    slots: [
+      {
+        start: 29912423,
+        end: null,
+        trackIds: cohort5TrackIds,
+        role: "guardian",
+      }, // cohort 5
     ],
   },
 ];